register xml handlers on the xml package
Until now the xml handlers were registered in the constructor of the
class handlers.Database (see __init__.py); now they are registered in
the init file of the xml package.

This avoids a circular import. But it changes the behaviour of the Database
class and sub-classes (i.e., it could break third party code).
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.4--patch-3