add the method _get_virtual_handler to folder handlers
Add the method "_get_virtual_handler(segment)" to folder handlers.
It is called by "get_handler" when the requested resource does not
exist, so we have a chance to do something "magic" and return a
"virtual" handler for a resource that lives somewhere else.

Before it was the "_get_handler(segment)" who had to deal with these
situations. Now it only must return a handler for resources that
exists; its new prototype is "_get_handler(segment, resource).

Beware, this API change is likely to break existing code.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.4--patch-26