handlers, change "load()" to "load(resource=None)"
Improve the handlers API. Now the "load" method accepts the parameter
resource, which specifies wich resource's data must be loaded, if
none is given the resource the handler is attached to will be used.

The method "_load" also has changed its prototype, now it expects
a resource as parameter. Note that any code that currently defines
"_load" must be changed.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.4--patch-65