catalog, fix unindexing
When a term was unindexed, the tree in memory (IIndex.Tree.children)
was cut if empty, but it was not cut on the resource. This inconsistency
between the resource and the handler lead to errors in further operations,
and finally to database inconsistency. Now the tree is not cut at all.

A side effect of the change is a performance improvement in re-indexing
operations.

There is another change, IDocument instances timestamp is kept up-to-date,
this brings further perfomance improvement.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.4--patch-85