handlers, implement the API for "the handler factory"
The design changes aimed to provide a better API to resolve the
question: for a given resource, which handler should be used?
have reached a milestone.

The API has been improved as documented in the section "The handler
factory" of the chapter "Resource Handlers". In particular the new
class method 'build_handler' is available:

  Handler.build_handler(resource)
 
It remains to finish the documentation about how to build your own
handlers, and to polish the code.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.5--patch-15