xhtml handler, change "get_messages" semantics
Change the semantics of the "get_messages" method from the handler
class "XHTML". Before it returned a set with all the messages found;
now it returns a list, with the same messages, and ordered by the
position they were found in the resource.

Update the unit tests within itools.xml
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.5--patch-22