optimize "Folder.get_handler" for outdated handlers
The method "Folder.get_handler" has been optimized for the
case where a handler is found in the cache but it is not
up-to-date.

Before the handler instance was discarded and a new call to
"self._get_handler" was made. This not only re-loaded the
resource, but also removed the handler instance to create a
new one.

Now the code just calls "handler.load", what re-loads the
resource, but keeps the same handler instance.

The performance has been improved by a factor of 2.5 (for the
case explained, and when the re-load weight is close to 0).

Also, the code readibility has been increased.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.6--patch-37