clean the folder's cache on "Folder.load" (fixes catalog bug)
Now the load operation for folders starts by cleaning the cache.

This has a very negative performance impact; for example, if you
create a new resource into a folder, the folder's modification
time will change, then it will be re-loaded the next time; if the
handler has a lot of content, it will be cleaned, even if it is
up-to-date.

We keep this code for now because it is safer. For example for the
catalog, if the cache is not cleaned the indexes won't be up-to-date,
hence causing errors.

Anyway, this code must be removed, another solution must be found
for the catalog in particular, and for all folders in general.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.6--patch-38