uri, now "str(get_reference(''))" returns "." instead of ""
The browsers behaviour for "" and "." links is different. This fixes
a problem in iKaaro, where the translated templates contained links
as "" while the original template was ".".

However, this may introduce other bugs...

The point is that the links ".", "" and "#" are the same thing from the
point of view of itools.uri; what must be done is to check what they
actually mean from the point of view of the W3C.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.6--patch-39