handlers, folder's state keeps resource names (in the cache)
With this patch when folder handler is loaded its cache initialized
with the resource names. This is to say, the handler state are the
the names of the resources it contains.

This change prevents to access the folder's resource on read operations,
hence improving the performance. How much depends on the speed of the
underlying resource layer.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.7--patch-10