handlers, the timestamp is initialized with its resource mtime
Now the handler timestamp is initialized with its resource's timestamp,
instead of with the current time. So now we are able to know wether a
handler has changed, i.e. the resource is outdated.

The method "has_changed" has been added to the API.

This change will allow performance improvements.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.7--patch-11