resources, fix the modification time of Zope 2 resources
Zope 2 resources should use "bobobase_modification_time()", so if a
resource is modified through the ZMI (for example) the resource is
reloaded the next time.

But unfortunately we can't use it.

Because the "bobobase_modification_time" is set with the ZODB transaction
commit, which happens always at the end, after the handler timestamp has
been set. This means that the resource modification time is always newer
than the handler's timestamp; so the resource always would be reloaded
after a modification, what is unacceptable.

This is the reason we keep ourselves the modification time, within the
persistent variable 'mtime'. When the resource is loaded for the first
time the on memory variable "bobo_time" is initialized; the modification
time of a resource is defined as the newer of the "object.mtime" and
"resource.bobo_time" variables.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.7--patch-36