packaging, fix setup.py, use relative paths (by Herve)
Using relative paths solves the problem found by Herve when building
Debian packages.

Note that there is a much better solution for Python 2.4 that should
be adopted once support for Python 2.3 is dropped, see:

  http://arch.oursours.net/zoom/herve@oursours.net--public/
    itools--herve--0.7--patch-6/setup.py

This change integrates the patchs #6, half of #8, #11, #12 and #13.
They have not been applied directly because the patch #8 was mixed
with another non related change and produced conflicts.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.7--patch-39