handlers, optimize CSV deserialization
Separate the parsing from the load. Now there is a generator
("CSV.parse") independent of the CSV handler class, which can
be used by specialized handler classes to load the resource
in one step.

Also, now "Text.guess_encoding" is a class method, so it can
be used without a "Text" instance.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.7--patch-64