handlers, rollback if there is an error in the commit phase
If an error happens when committing a transaction, it is rolled back.
This way the handlers state is always consistent.

However, the resource layer may still be left into an inconsistent
state (there is no problem with "itools.resources.zope2" because it
provides transactions on its own, but there is problem with the
filesystem resource layer for instance).

Also, added a missing import.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.8--patch-71