handlers, split new class Node from Handler
Start to separate the category "handler as a logical representation of
a resource" from "handler as a node in the handler tree".

For now the implementation change has been to split the "handler tree"
logic from the Handler class into a new class named "Node". This is a
base class and Handler inherits from it.

This change allows, for example, to traverse files. The CSV handler
has been modified to illustrate this new ability, now it is possible
to do "csv_handler.get_handler('5')" to get a "handler" for the fifth
row of the CSV file.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.9--patch-60