html, omit entity references not in the source encoding
Entity references allow to use characters that can not be represented
on the document encoding. But our policy is to translate these entity
references to that encoding, what is a contradiction. Just try to parse
a document on latin with the entity "&rsquo;".

For now, we ignore these entities. This needs to be fixed.
git-archimport-id: jdavid@itaapy.com--public/itools--main--0.9--patch-104