[uri] Now path segments are stored as strings, the name and parameters are
calculated when needed. This is to aknowledge the fact that we rarely use
the parameters.

The purpose is to simplify the API (we will write just "segment" instead
of "segment.name" if we don't care about the parameters), and to reduce
the memory and CPU usage.

Also, now we have "n" parameters, not just one (or none). As the RFC says.
So it will be "segment.params" instead of "segment.param", and the returned
value will be a list of byte strings instead of only one string.