csv: allow (optionally) more chars in param values
This change is done for the metadata in itools.database, not used by
ical. One day we will: (1) remove the table handler (2) have a separate
parser for ical and for the metadata.

If the datatype of a paramter has defined the 'escape' attribute to
true, some characters will be allowed (wich otherwise are not): cr, lf,
and double quotes.