Database: remove "sync_filesystem"
sync_filesystem was use to check if handlers were modified
in the database (via vim for example) so we can update cache.
We remove this verification since it's take IO time
Moreover the database state was not updated.
Now we can use ikaaro rest API to do this kind of changes on DB.