String Datatype is now str (unicode)
As far as itools.datatypes is concerned, we don't have to deal with
bytes anymore, because in Python 3 we're getting unicode strings
(itools.datatypes is only used for text files).

We still have the String and Unicode datatypes, where the behaviour
of Unicode has not changed. The difference between them is:

- String.default is None, Unicode.default is the empty string
- The catalog will index a String field as is, but will split a Unicode
  field in words.

What has changed spicifically is that now String decodes/encodes from/to
str instead of bytes.