Don't call make_room so often
Optimize the case where many resources are created before saving.

The problem is the cache is full with dirty handlers, which cannot be
removed from the cache.

A better (and more complex) solution would be to keep the dirty handlers
apart: do not consider them at all when looking for candidate handlers
to discard.